dataset/requirements.txt
numpy~=1.24.3
matplotlib~=3.7.1
tqdm~=4.65.0
biopython
more-itertools
pyensembl~=2.2.8
requests~=2.30.0
scikit-bio
pandas~=2.0.1
tables
ushuffle